+++ allennlp_models/rc/models/transformer_qa.py
@@ -1,18 +1,20 @@
 import logging
+import warnings
 from typing import Any, Dict, List, Optional, Tuple
 
 import numpy as np
 import torch
-from allennlp.common.util import sanitize_wordpiece
+import torch.distributed as dist
+from torch import nn
+from torch.nn.functional import cross_entropy
+
+from allennlp.common.util import sanitize_wordpiece, is_distributed
 from allennlp.modules.text_field_embedders import BasicTextFieldEmbedder
 from allennlp.modules.token_embedders import PretrainedTransformerEmbedder
 from allennlp.nn.util import get_token_ids_from_text_field_tensors
-from torch import nn
-
 from allennlp.data import Vocabulary
 from allennlp.models.model import Model
 from allennlp.training.metrics import BooleanAccuracy, CategoricalAccuracy
-from torch.nn.functional import cross_entropy
 
 from allennlp_models.rc.models.utils import (
     get_best_span,
@@ -247,9 +249,30 @@ class TransformerQA(Model):
         """
         _best_spans = best_spans.detach().cpu().numpy()
 
+        min_node_batch_size = len(metadata)
+        if is_distributed():
+            # In the distributed case, if the batch isn't balanced across nodes, we'll get a dead-lock
+            # if one node has more instances in a batch than another node.
+            if dist.get_backend() == "nccl":
+                device = torch.cuda.current_device()
+            else:
+                device = torch.device("cpu")
+            _min_node_batch_size = torch.tensor(min_node_batch_size, dtype=torch.int, device=device)
+            dist.all_reduce(_min_node_batch_size, op=dist.ReduceOp.MIN)
+            min_node_batch_size = _min_node_batch_size.item()
+
+            if min_node_batch_size != len(metadata):
+                warnings.warn(
+                    "Nodes have different batch sizes! Some instances will be excluded from metric computation.",
+                    UserWarning,
+                )
+
         best_span_strings = []
         for (metadata_entry, best_span, cspan, cls_ind) in zip(
-            metadata, _best_spans, context_span, cls_index or (0 for _ in range(len(metadata)))
+            metadata,
+            _best_spans,
+            context_span,
+            cls_index or (0 for _ in range(min_node_batch_size)),
         ):
             context_tokens_for_question = metadata_entry["context_tokens"]