+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - 'master' branch renamed to 'main'
 
+### Fixed
+
+- Fixed the potentially for a dead-lock when training the `TransformerQA` model on multiple GPUs
+  when nodes receive different sized batches.
+
 ## [v1.3.0](https://github.com/allenai/allennlp-models/releases/tag/v1.2.2) - 2020-12-15
 
 ### Fixed